# db: reduce number of iterator allocations when read amp is high