--- a/db_test.go
+++ b/db_test.go
@@ -1170,6 +1170,39 @@ func BenchmarkDelete(b *testing.B) {
 	})
 }
 
+func BenchmarkNewIterReadAmp(b *testing.B) {
+	for _, readAmp := range []int{10, 100, 1000} {
+		b.Run(strconv.Itoa(readAmp), func(b *testing.B) {
+			opts := &Options{
+				FS:                    vfs.NewMem(),
+				L0StopWritesThreshold: 1000,
+			}
+			opts.private.disableAutomaticCompactions = true
+
+			d, err := Open("", opts)
+			require.NoError(b, err)
+
+			for i := 0; i < readAmp; i++ {
+				require.NoError(b, d.Set([]byte("a"), []byte("b"), NoSync))
+				require.NoError(b, d.Flush())
+			}
+
+			require.Equal(b, d.Metrics().ReadAmp(), readAmp)
+
+			b.StopTimer()
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				b.StartTimer()
+				iter := d.NewIter(nil)
+				b.StopTimer()
+				require.NoError(b, iter.Close())
+			}
+
+			require.NoError(b, d.Close())
+		})
+	}
+}
+
 func verifyGet(t *testing.T, r Reader, key, expected []byte) {
 	val, closer, err := r.Get(key)
 	require.NoError(t, err)