This patch reduces the number of slice allocations for a new
iterator when the read amplification is larger than 10.

Pebble benchmarks are mostly unchanged:
```
name                old ops/sec  new ops/sec  delta
ycsb/A/values=64      819k ± 6%    812k ± 7%    ~     (p=0.739 n=10+10)
ycsb/B/values=64     1.09M ± 7%   1.08M ± 5%    ~     (p=0.853 n=10+10)
ycsb/C/values=64     2.27M ±28%   2.40M ±22%    ~     (p=0.529 n=10+10)
ycsb/D/values=64      897k ± 7%    891k ± 7%    ~     (p=0.971 n=10+10)
ycsb/E/values=64      214k ± 5%    213k ± 6%    ~     (p=0.739 n=10+10)
ycsb/F/values=64      239k ± 1%    239k ± 1%    ~     (p=0.684 n=10+10)
ycsb/A/values=1024   77.1k ± 4%   77.4k ± 3%    ~     (p=0.796 n=10+10)
ycsb/B/values=1024    323k ± 8%    326k ± 9%    ~     (p=0.912 n=10+10)
ycsb/C/values=1024    773k ±22%    766k ±13%    ~     (p=0.400 n=10+9)
ycsb/D/values=1024    164k ±10%    168k ±11%    ~     (p=0.436 n=10+10)
ycsb/E/values=1024   53.0k ±10%   53.2k ±10%    ~     (p=0.912 n=10+10)
ycsb/F/values=1024   26.5k ± 1%   26.5k ± 2%    ~     (p=0.971 n=10+10)

name                old read     new read     delta
ycsb/A/values=64     65.8G ± 1%   65.8G ± 1%    ~     (p=0.481 n=8+9)
ycsb/B/values=64     37.4G ±10%   35.9G ±20%    ~     (p=0.481 n=10+10)
ycsb/C/values=64     2.80G ± 6%   2.84G ± 2%    ~     (p=0.173 n=10+8)
ycsb/D/values=64     37.4G ± 6%   38.1G ±10%    ~     (p=0.481 n=10+10)
ycsb/E/values=64     14.8G ±15%   14.7G ±12%    ~     (p=0.720 n=9+10)
ycsb/F/values=64      162G ± 2%    162G ± 2%    ~     (p=0.853 n=10+10)
ycsb/A/values=1024    166G ± 2%    167G ± 3%    ~     (p=0.436 n=10+10)
ycsb/B/values=1024    109G ± 5%    110G ± 5%    ~     (p=0.579 n=10+10)
ycsb/C/values=1024   62.3G ± 4%   63.1G ± 3%    ~     (p=0.143 n=10+10)
ycsb/D/values=1024    112G ± 7%    113G ± 8%    ~     (p=0.631 n=10+10)
ycsb/E/values=1024   63.9G ± 4%   64.0G ± 4%    ~     (p=0.971 n=10+10)
ycsb/F/values=1024    169G ± 1%    169G ± 0%    ~     (p=0.353 n=10+10)

name                old write    new write    delta
ycsb/A/values=64     96.0G ± 1%   95.9G ± 1%    ~     (p=0.963 n=8+9)
ycsb/B/values=64     41.5G ± 9%   40.0G ±18%    ~     (p=0.481 n=10+10)
ycsb/C/values=64     2.71G ± 6%   2.75G ± 2%    ~     (p=0.173 n=10+8)
ycsb/D/values=64     43.0G ± 6%   43.6G ±10%    ~     (p=0.481 n=10+10)
ycsb/E/values=64     16.3G ±16%   15.9G ±11%    ~     (p=0.481 n=10+10)
ycsb/F/values=64      192G ± 2%    192G ± 1%    ~     (p=0.971 n=10+10)
ycsb/A/values=1024    193G ± 2%    195G ± 3%    ~     (p=0.481 n=10+10)
ycsb/B/values=1024    121G ± 5%    123G ± 5%    ~     (p=0.579 n=10+10)
ycsb/C/values=1024   62.3G ± 4%   63.0G ± 3%    ~     (p=0.143 n=10+10)
ycsb/D/values=1024    122G ± 7%    124G ± 8%    ~     (p=0.631 n=10+10)
ycsb/E/values=1024   67.2G ± 4%   67.4G ± 4%    ~     (p=0.971 n=10+10)
ycsb/F/values=1024    203G ± 1%    203G ± 0%    ~     (p=0.353 n=10+10)

name                old r-amp    new r-amp    delta
ycsb/A/values=64      6.65 ± 2%    6.64 ± 2%    ~     (p=0.699 n=10+10)
ycsb/B/values=64      3.76 ± 3%    3.72 ± 1%  -1.15%  (p=0.044 n=10+9)
ycsb/C/values=64      1.68 ±30%    1.75 ±33%    ~     (p=0.562 n=10+10)
ycsb/D/values=64      4.18 ± 1%    4.18 ± 1%    ~     (p=0.861 n=10+10)
ycsb/E/values=64      3.18 ± 6%    3.16 ± 4%    ~     (p=0.827 n=10+9)
ycsb/F/values=64      0.00         0.00         ~     (all equal)
ycsb/A/values=1024    20.5 ±16%    19.8 ±17%    ~     (p=0.529 n=10+10)
ycsb/B/values=1024    7.58 ± 2%    7.61 ± 2%    ~     (p=0.425 n=10+10)
ycsb/C/values=1024    2.96 ± 5%    3.00 ± 2%    ~     (p=0.076 n=9+7)
ycsb/D/values=1024    10.3 ± 5%    10.2 ± 5%    ~     (p=0.197 n=10+10)
ycsb/E/values=1024    16.8 ± 8%    16.6 ± 7%    ~     (p=0.591 n=10+10)
ycsb/F/values=1024    0.00         0.00         ~     (all equal)

name                old w-amp    new w-amp    delta
ycsb/A/values=64      3.21 ± 2%    3.21 ± 2%    ~     (p=0.789 n=10+10)
ycsb/B/values=64      10.3 ±10%    10.0 ±17%    ~     (p=0.579 n=10+10)
ycsb/C/values=64      0.00         0.00         ~     (all equal)
ycsb/D/values=64      12.9 ± 3%    13.2 ± 4%    ~     (p=0.055 n=10+10)
ycsb/E/values=64      21.0 ± 9%    20.2 ±12%    ~     (p=0.211 n=9+10)
ycsb/F/values=64      10.8 ± 1%    10.8 ± 0%    ~     (p=0.640 n=10+10)
ycsb/A/values=1024    7.65 ± 2%    7.72 ± 1%    ~     (p=0.209 n=10+8)
ycsb/B/values=1024    11.6 ± 4%    11.6 ± 3%    ~     (p=0.988 n=10+10)
ycsb/C/values=1024    0.00         0.00         ~     (all equal)
ycsb/D/values=1024    22.9 ± 4%    22.7 ± 4%    ~     (p=0.343 n=10+10)
ycsb/E/values=1024    39.1 ± 8%    39.0 ± 8%    ~     (p=0.796 n=10+10)
ycsb/F/values=1024    11.8 ± 1%    11.8 ± 1%    ~     (p=0.696 n=10+10)
```

In a microbenchmark specifically benchmarking DB.NewIter,
I observed no change in performance at small read amps
but a ~35% time improvement for larger read amps. I also
observed a ~35% bytes allocation reduction and a ~98%
number of allocations reduction for large read amps.
```
name                    old time/op    new time/op    delta
NewIterReadAmp/10-16      1.45µs ±18%    1.37µs ±22%     ~     (p=0.247 n=10+10)
NewIterReadAmp/100-16     31.7µs ± 9%    21.2µs ± 1%  -33.28%  (p=0.000 n=9+9)
NewIterReadAmp/1000-16     346µs ± 6%     228µs ± 1%  -34.10%  (p=0.000 n=9+8)

name                    old alloc/op   new alloc/op   delta
NewIterReadAmp/10-16       0.00B          0.00B          ~     (all equal)
NewIterReadAmp/100-16      105kB ± 0%      68kB ± 0%  -35.40%  (p=0.000 n=10+10)
NewIterReadAmp/1000-16    1.00MB ± 0%    0.64MB ± 0%  -35.74%  (p=0.000 n=10+10)

name                    old allocs/op  new allocs/op  delta
NewIterReadAmp/10-16        0.00           0.00          ~     (all equal)
NewIterReadAmp/100-16       94.0 ± 0%       2.0 ± 0%  -97.87%  (p=0.000 n=10+10)
NewIterReadAmp/1000-16     1.00k ± 0%     0.00k ± 0%  -99.80%  (p=0.000 n=10+10)
```
Fixes https://github.com/cockroachdb/pebble/issues/1122.