@andyyang890 2021-05-27
These are the benchstat results when I run the microbenchmark on my computer:
```
name        old time/op    new time/op    delta
NewIter-16    33.7µs ± 4%    29.4µs ± 4%  -12.78%  (p=0.000 n=13+13)

name        old alloc/op   new alloc/op   delta
NewIter-16     105kB ± 0%      68kB ± 0%  -35.41%  (p=0.000 n=13+13)

name        old allocs/op  new allocs/op  delta
NewIter-16      94.0 ± 0%       2.0 ± 0%  -97.87%  (p=0.000 n=13+13)
```